drivers: Receive Reader instead of buffer on save
We want to use this lib for saving large files to ObjectStore
as well for the VOD tasks. Since they can be really large, we
don't want to have to buffer the whole files in memory for
saving. So changing this API to receive a Reader instead makes
more sense.